backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
motor==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
groq==0.4.1
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
bcrypt==4.1.2
